components out of styles
