+++ agency-dashboard/src/DashboardView.tsx
@@ -15,20 +15,25 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
 
+import { ReactComponent as DownloadIcon } from "@justice-counts/common/assets/download-icon.svg";
+import { ReactComponent as GridIcon } from "@justice-counts/common/assets/grid-icon.svg";
+import { ReactComponent as InfoIcon } from "@justice-counts/common/assets/info-icon.svg";
+import { ReactComponent as ShareIcon } from "@justice-counts/common/assets/share-icon.svg";
 import { DatapointsView } from "@justice-counts/common/components/DataViz/DatapointsView";
 import { observer } from "mobx-react-lite";
 import React, { useEffect } from "react";
 import { useLocation, useNavigate, useParams } from "react-router-dom";
 
 import {
-  AllMetricsButton,
+  AllMetricsButtonContainer,
+  AllMetricsButtonCountContainer,
+  AllMetricsButtonText,
   Container,
   LeftPanel,
   LeftPanelBackButton,
-  MetricOverviewActionDownloadButton,
-  MetricOverviewActionInfoButton,
+  MetricOverviewActionButtonContainer,
+  MetricOverviewActionButtonText,
   MetricOverviewActionsContainer,
-  MetricOverviewActionShareButton,
   MetricOverviewContent,
   MetricOverviewTitle,
   MetricTitle,
@@ -37,6 +42,44 @@ import {
 } from "./DashboardView.styles";
 import { HeaderBar } from "./Header/HeaderBar";
 import { useStore } from "./stores";
+
+export const MetricOverviewActionShareButton = () => (
+  <MetricOverviewActionButtonContainer>
+    <ShareIcon />
+    <MetricOverviewActionButtonText>Share</MetricOverviewActionButtonText>
+  </MetricOverviewActionButtonContainer>
+);
+
+export const MetricOverviewActionDownloadButton = () => (
+  <MetricOverviewActionButtonContainer>
+    <DownloadIcon />
+    <MetricOverviewActionButtonText>
+      Download Data
+    </MetricOverviewActionButtonText>
+  </MetricOverviewActionButtonContainer>
+);
+
+export const MetricOverviewActionInfoButton = () => (
+  <MetricOverviewActionButtonContainer>
+    <InfoIcon />
+    <MetricOverviewActionButtonText>Learn More</MetricOverviewActionButtonText>
+  </MetricOverviewActionButtonContainer>
+);
+
+export const AllMetricsButton = ({
+  metricsCount,
+}: {
+  metricsCount: number;
+}) => (
+  <AllMetricsButtonContainer>
+    <GridIcon />
+    <AllMetricsButtonText>All Metrics</AllMetricsButtonText>
+
+    <AllMetricsButtonCountContainer>
+      {metricsCount}
+    </AllMetricsButtonCountContainer>
+  </AllMetricsButtonContainer>
+);
 
 const DashboardView = () => {
   const navigate = useNavigate();
